Improve DevMenu (#120)
